test(gateway): give identity-conflict coverage its own file
Both lifecycle suites crossed the 1000-line test cap once this PR's lineage cases
landed in them. The cases that moved are the ones this change owns — the delete
successor table and the archive preparation race — so they now sit together in
server.sessions.identity-conflict.test.ts instead of behind a suppression.
